requirements.txt
discord.py
gspread
dotenv